requirements.txt
requests
msgpack
kafka-python
psycopg2-binary
fire
